Fix complex-valued wave example.

+++ examples/wave/wave.py
@@ -28,7 +28,6 @@ from hedge.mesh import TAG_ALL, TAG_NONE
 def main(write_output=True, 
         dir_tag=TAG_NONE, neu_tag=TAG_NONE, rad_tag=TAG_ALL, 
         flux_type_arg="upwind", dtype=numpy.float64, debug=[]):
-    from hedge.timestep import RK4TimeStepper
     from pytools.stopwatch import Job
     from math import sin, cos, pi, exp, sqrt
 
@@ -59,7 +58,8 @@ def main(write_output=True,
         mesh_data = rcon.receive_mesh()
 
     discr = rcon.make_discretization(mesh_data, order=4, debug=debug)
-    stepper = RK4TimeStepper()
+    from hedge.timestep import RK4TimeStepper
+    stepper = RK4TimeStepper(dtype=dtype)
 
     from hedge.visualization import VtkVisualizer
     if write_output:
